requirements.txt
joblib>=1.2
numpy>=1.23
pandas>=1.5
PyYAML>=6.0
scikit-learn>=1.2
pytest>=7.2
